move `buildObjectTree` to `query.svelte`
